v9 làm riêng cho 20251110_Culture từ chuỗi tiền n+1 phải thắng tổng 2 lần các mức tiền chuỗi tiền n thì quay về n

+++ ABXHubSolution/XocDiaSoiLiveKH24/Tasks/Top10PatternFollowTask.cs
@@ -186,7 +186,7 @@ namespace XocDiaSoiLiveKH24.Tasks
                     // cần biến local để truyền ref
                     int chainIndex = ctx.MoneyChainIndex;
                     int chainStep = ctx.MoneyChainStep;
-                    long chainProfit = ctx.MoneyChainProfit;
+                    double chainProfit = ctx.MoneyChainProfit;
 
                     MoneyHelper.UpdateAfterRoundMultiChain(
                         ctx.StakeChains,
