+++ LamRieng/20251110_Culture từ chuỗi tiền n+1 phải thắng tổng 2 lần các mức tiền chuỗi tiền n thì quay về n/XocDiaLiveHit/Tasks/StateTransitionBiasTask.cs
@@ -77,7 +77,7 @@ namespace XocDiaLiveHit.Tasks
                     // cần biến local để truyền ref
                     int chainIndex = ctx.MoneyChainIndex;
                     int chainStep = ctx.MoneyChainStep;
-                    long chainProfit = ctx.MoneyChainProfit;
+                    double chainProfit = ctx.MoneyChainProfit;
 
                     MoneyHelper.UpdateAfterRoundMultiChain(
                         ctx.StakeChains,
